requirements.txt
Django==4.0
environs==9.5.0
djangorestframework==3.14.0
psycopg2-binary==2.9.3
geopy==2.2.0
dj-database-url==0.5.0
django-debug-toolbar==4.1.0
